LayoutTests/platform/gtk/media/nodesFromRect-shadowContent-expected.txt

PASS All correct nodes found for rect
PASS All correct nodes found for rect
This test only runs in DRT!

EXPECTED (video.controls != 'null') OK
EVENT(canplaythrough)
END OF TEST
